xgen/requirements.txt
panda
torch
gradio
numpy
transformers
tiktoken